backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.9
aiohttp==3.9.5

# AI Services
google-generativeai==0.8.0
sentence-transformers==5.2.0

# Database - Supabase (User Data)
supabase==2.10.0

# Database - Turso (Drug Data)
libsql-client

# Database - Qdrant (Vector Embeddings)
qdrant-client==1.7.0

# Marketplace
razorpay==1.4.2
agora-token-builder==1.0.0

# Security
slowapi==0.1.9
curl_cffi

# Web Scraping
playwright==1.45.0
requests==2.32.3
redis==5.0.8

lxml==5.2.2

python-socketio
